server/requirements.txt
flask
flask-cors
pyjwt
python-dotenv
psycopg2-binary
argon2-cffi
zmq
ta
pandas
stable_baselines3
stable-baselines3[extra]
gymnasium[box2d]
torch
schedule
requests
flasgger